Add /system/info endpoint for faas-swarm
The purpose of this change is to make faas-swarm compatible with the
latest version of the API Gateway in OpenFaaS. Each provider should
be able to advertise the underlying provider and the version of the
code.

A new version of faas-provider (0.6) is vendored into the repo
which brings in an optional health endpoint (turned off by default).

The version and SHA is being passed to the provider from server.go.
Dockerfiles are updated to inject version and SHA string at build time in
version.go.

A unit test has been added for the /system/info endpoint.

Signed-off-by: Vivek Singh <vivekkmr45@yahoo.in>