Add CPU limits for Swarm using NanoCPU
CPU limits and requests/reservations can now be specified in
NanoCPU units. NanoCPU units are cryptic but explained briefly
in the following GitHub issue [0] and moby code [1]

Tested to ensure that memory and CPU can be given together
and separtely if only one needs to be limited. Copies my work
done on faas-netes several months ago.

[0] https://github.com/moby/moby/issues/30222
[1] https://github.com/moby/moby/blob/e4a0dbc47232e3a9da4cfe6ce44f250e6e85ed43/daemon/cluster/executor/container/container.go#L328-L332

Signed-off-by: Alex Ellis (VMware) <alexellis2@gmail.com>