Fix deploy to use valid docker image names including tags
This resolves the issue #19 that throws an error when using
a tag on the image name.

There are several variations to a valid docker image name.
Repository may, or may not, be included. Default is docker.io.
Username may, or may not, be included.
Repository name is always included.
An image tag may, or may not, be included. Default is 'latest'.

This update takes all those scenarios into account when getting
the authentication credentials to allow private functions
to be deployed from docker hub, or another repository

Signed-off-by: Burton Rheutan <rheutan7@gmail.com>