Fix .travis.yml to address "Could not parse .travis.yml" message
The travis build stopped working on 26/03.  There were 3 commits on this day.
The first two saw the build run green, it was after the third that the error
message of "Could not parse .travis.yml" was observed.  The line added in the
3rd commit on this day was to echo the $TRAVIS_TAG.  This line is linting as
it breaks the yaml format.

Offending line removed to regress to previous working state

Signed-off-by: rgee0 <richard@technologee.co.uk>