Added tests to secrets handler
To test that endpoint I changed also functions parameters
from `*client.Client` to `client.SecretAPIClient` which is
interface created by docker for that purposes.
With that in place, I could create fake dockerClient
and pass it to `MakeSecretsHandler` function and then test it.

I also fixed a small bug, which was I found because of tests.
In updateSecret function I was checking if updateSecretErr
is not nil, and then instead of return this error wrapped
by new error, I was returning getSecretErr which was nil,
so function was panicking.

Signed-off-by: Bart Smykla <bsmykla@vmware.com>